requirements.txt
pyserial==3.5
tk==0.1.0 